place dataSource in a separate file
